adding grass, road, car squares
the object is 10x10 if it ends in 10, 15x1 if it ends ins 15, and 20x20
if it ends in 20.
I made the road2-10, road2-20 texture. It's still a bit meh in my
opinion but we wouldn't have to worry about copyright. I'll work on a
grass one and maybe a better road one. In my opinion our car won't look
like a real car unless we have 20x20 squares.